ci(canon): require tsgo for Rust tests
